// apps/avto-shop/src/libs/enums/member.enum.ts
export enum Type {
    USER = 'USER',
    AGENT = 'AGENT',
    ADMIN = 'ADMIN',
    DEALER = 'DEALER',
    SERVICE = 'SERVICE',
    SELLER = 'SELLER',
};